Add confirm

+++ src/components/dashboard/posts/DeletePostButton.js
@@ -12,11 +12,15 @@ export default function DeletePostButton({ id }) {
 	const url = `/wp/v2/posts/${id}`;
 
 	async function handleDelete() {
-		try {
-			await http.delete(url);
-			history.push("/dashboard/posts");
-		} catch (error) {
-			setError(error);
+		const confirmDelete = window.confirm("Delete this post?");
+
+		if (confirmDelete) {
+			try {
+				await http.delete(url);
+				history.push("/dashboard/posts");
+			} catch (error) {
+				setError(error);
+			}
 		}
 	}
